rebulid player state behaviors
