+++ Silksong/Assets/Scripts/Player/PlayerController/PlayerStateActions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e3558255c21667f46afbd6e0535d33b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
