+++ Silksong/Assets/Scripts/Player/PlayerController/PlayerStateActionsForCat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e1320929c78cd34898a1e0e49c6107b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
