+++ Silksong/Assets/Scripts/Player/PlayerController/PlayerStateActionsUnderWater.cs
@@ -1,0 +1,80 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PlayerSwim : PlayerAction
+{
+    public PlayerSwim(PlayerController playerController) : base(playerController) { }
+
+    public override void StateUpdate()
+    {
+        playerController.SwimUnderWater();
+    }
+    public override void StateStart(EPlayerState oldState)
+    {
+        playerController.IsUnderWater = true;
+        //入水后0.2s内禁止向上
+        //setRigidGravityScale(playerInfo.normalGravityScale/2);
+    }
+    public override void StateEnd(EPlayerState newState)
+    {
+        playerController.m_Transform.localRotation = Quaternion.Euler(0, 0, 0);
+        playerController.IsUnderWater = false;
+        //setRigidGravityScaleToNormal();
+    }
+}
+
+
+public class PlayerSprintInWater : PlayerAction
+{
+    public PlayerSprintInWater(PlayerController playerController) : base(playerController) { }
+    private bool sprintReady;
+    public bool SprintReady
+    {
+        get { return sprintReady; }
+        set
+        {
+            sprintReady = value;
+            playerController.PlayerAnimator.SetBool(playerController.animatorParamsMapping.SprintReadyParamHash, sprintReady);
+        }
+    }
+
+    public override void StateStart(EPlayerState oldState)
+    {
+        playerController.setRigidGravityScale(0);
+        int x = playerController.playerInfo.playerFacingRight ? 1 : -1;
+        int y = 0;
+        if (PlayerInput.Instance.vertical.Value != 0)
+        {
+            y = PlayerInput.Instance.vertical.Value > 0 ? 1 : -1;
+        }
+        playerController.setRigidVelocity(new Vector2(playerController.playerInfo.sprintSpeed * x, playerController.playerInfo.sprintSpeed * y));
+        playerController.gravityLock = true;
+        playerController.IsUnderWater = false;
+    }
+
+    public override void StateEnd(EPlayerState newState)
+    {
+        Debug.Log(playerController.IsUnderWater);
+        playerController.gravityLock = false;
+        if (!playerController.IsUnderWater)
+        {
+            //冲出水面重力正常
+            playerController.setRigidGravityScaleToNormal();
+        }
+        else
+        {
+            playerController.setRigidGravityScale(playerController.playerInfo.gravityUnderWater);
+        }
+        playerController.setRigidVelocity(Vector2.zero);
+        playerController.StartCoroutine(sprintCdCount());
+    }
+
+
+    public IEnumerator sprintCdCount()
+    {
+        SprintReady = false;
+        yield return new WaitForSeconds(Constants.SprintCd);
+        SprintReady = true;
+    }
+}
