+++ Silksong/Assets/Scripts/Player/PlayerController/PlayerStateActionsUnderWater.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 26b30f387ba94c34b946c6beb6aa0171
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
